HF component update to not use fsspec components
Pull Request resolved: https://github.com/pytorch/pytorch/pull/159405

Update HF components to not inherit from fsspec components and instead use filesystem writer/reader. The reason is because there doesn't seem to be much of a need for fsspec, since users are using mounted storage. Using local storage will allow for performance improvements because we can take advantage of the safe_open API provided by HF safetensors, which is signifcant performance wins over reading bytes and converting to tensors which is what we are doing now.
ghstack-source-id: 300168979
@exported-using-ghexport

Differential Revision: [D78993550](https://our.internmc.facebook.com/intern/diff/D78993550/)